feat: add homepage
Add very basic homepage.
In line 97 in index earlier i added some code by mistake and it was making problems so i removed it.
