+++ src/components/menu.tsx
@@ -1,4 +1,7 @@
+import { GoHome } from "react-icons/go";
 import { MdAdd } from "react-icons/md";
+
+import Link from "next/link";
 
 import { Button } from "@/components/button";
 import { Drawer } from "@/components/drawer";
@@ -20,6 +23,12 @@ export const MenuDrawer = ({
   return (
     <Drawer open={open}>
       <div className="flex flex-col justify-center gap-8 py-4 pb-32 ">
+        <Button
+          className="mx-auto w-3/5"
+          icon={<GoHome className="h-5 w-5 fill-current" />}
+        >
+          <Link href="/"> Back to Homepage</Link>
+        </Button>
         <Button
           className="mx-auto w-3/5"
           icon={<MdAdd />}
