+++ src/components/navbar.tsx
@@ -7,7 +7,7 @@ export type NavbarProps = HTMLAttributes<HTMLElement>;
 export const Navbar = ({ className, ...props }: NavbarProps) => (
   <nav
     className={clsx(
-      "fixed top-0 left-0 right-0 z-navbar flex justify-between gap-8 bg-black/50 px-4 backdrop-blur-xl",
+      "fixed top-0 left-0 right-0 z-navbar flex gap-8 bg-black/50 px-4 backdrop-blur-xl",
       className
     )}
     {...props}
